# Optimize Decimal calculations for 32-bit platforms