--- a/src/System.Private.CoreLib/src/System/Decimal.cs
+++ b/src/System.Private.CoreLib/src/System/Decimal.cs
@@ -335,7 +335,7 @@ namespace System
         //
         public static decimal Add(decimal d1, decimal d2)
         {
-            DecCalc.VarDecAdd(ref d1, ref d2);
+            DecCalc.DecAddSub(ref d1, ref d2, false);
             return d1;
         }
 
@@ -659,7 +659,7 @@ namespace System
         //
         public static decimal Subtract(decimal d1, decimal d2)
         {
-            DecCalc.VarDecSub(ref d1, ref d2);
+            DecCalc.DecAddSub(ref d1, ref d2, true);
             return d1;
         }