Contributes to https://github.com/dotnet/coreclr/issues/18249.
After these changes only multiplication is slower (1.15x) than the current native code. I've got a version using SSE2 intrinsics that's 1.09x slower than native, but I think it's not worth the additional complexity.
Below are the x86 stats (on x64 this change has only some minor improvements).

|  CMP |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
|  Native | 16.72 ns | 0.0208 ns | 0.0074 ns |   1.00 |
| CoreRT2 | 14.52 ns | 0.0186 ns | 0.0067 ns |   0.87 |

|  ADD |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
|  Native | 28.54 ns | 0.1237 ns | 0.0441 ns |   1.00 |
| CoreRT2 | 28.20 ns | 0.1221 ns | 0.0435 ns |   0.99 |

|  MUL |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
|  Native | 19.60 ns | 0.9202 ns | 0.3282 ns |   1.00 |
| CoreRT2 | 22.17 ns | 0.0437 ns | 0.0156 ns |   1.15 |

|  DIV |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
|  Native | 116.7 ns | 0.1891 ns | 0.0674 ns |   1.00 |
| CoreRT2 | 106.7 ns | 0.1221 ns | 0.0435 ns |   0.91 |