@jkotas 2018-07-11
> After these changes only multiplication is slower (1.15x) than the current native code

If this is hard to fix, I do not think this one-off regression would be a problem for integrating the managed decimal to CoreCLR.